Refactor: move styling to seperate file
