+++ src/App.js
@@ -1,9 +1,11 @@
-import InfoSection from "./components/InfoSection/InfoSection"
+// import InfoSection from "./components/InfoSection/InfoSection"
+import Dashboard from './components/Dashboard/Dashboard.jsx'
 
 function App() {
 	return (
 		<div>
-			<InfoSection/>
+			{/* <InfoSection/> */}
+			<Dashboard />
 		</div>
 	)
 }
