+++ src/components/Dashboard/Dashboard.jsx
@@ -4,7 +4,6 @@ import Grid from "@material-ui/core/Grid";
 import Button from "@material-ui/core/Button";
 import { Avatar } from "@material-ui/core";
 import "./styles.css";
-import { makeStyles } from "@material-ui/core/styles";
 import Post from "./Post/Post";
 
 import React from "react";
@@ -15,6 +14,7 @@ import Cities from "./Widgets/Cities";
 import Filters from "./Widgets/Filters";
 import SearchBar from "material-ui-search-bar";
 // import LinearProgress from "@material-ui/core/LinearProgress";
+import { useStyleDashboard } from "./Widgets/Style";
 
 export default function Dashboard() {
   const [state, setState] = React.useState({
@@ -43,32 +43,7 @@ export default function Dashboard() {
     window.addEventListener("resize", () => setResponsiveness());
   }, []);
 
-  const useStyles = makeStyles(() => ({
-    button: {
-      color: "white",
-      margin: "0 5px",
-      backgroundColor: "#3AAFA9",
-      "&:hover": {
-        backgroundColor: "#27938d",
-      },
-    },
-    widget: {
-      marginTop: "20px",
-      display: "flex",
-      justifyContent: "center",
-      height: "fit-content",
-    },
-    container: {
-      marginTop: "50px",
-    },
-    searchfield: {
-      height: "40px",
-      backgroundColor: "#f7f7f7",
-      boxShadow: "none",
-      border: "1px solid rgb(231, 231, 231)",
-    },
-  }));
-  const { container, searchfield, button, widget } = useStyles();
+  const { container, searchfield, button, widget } = useStyleDashboard();
   return (
     <div>
       <Navbar />
