+++ src/components/Dashboard/Widgets/Cities.jsx
@@ -1,45 +1,11 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
+// import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
 // import Avatar from "@material-ui/core/Avatar";
-import { grey } from "@material-ui/core/colors";
 import Button from "@material-ui/core/Button";
 // import ghost from "../../../Assets/Ghost.jpg";
-
-const useStyles = makeStyles(() => ({
-  root: {
-    maxWidth: 320,
-    marginBottom: "15px",
-    borderRadius: "20px",
-    width: "300px",
-  },
-  buttons: {
-    display: "flex",
-    flexWrap: "wrap",
-    justifyContent: "center",
-    margin: "15px 0px",
-  },
-  button: {
-    backgroundColor: "#3AAFA9",
-    color: "white",
-    width: "130px",
-    margin: "5px",
-    "&:hover": {
-      backgroundColor: "#27938d",
-    },
-  },
-  avatar: {
-    backgroundColor: grey[500],
-    width: "30px",
-    height: "30px",
-  },
-  header: {
-    backgroundColor: "#3AAFA9",
-    color: "white",
-    height: "15px",
-  },
-}));
+import { useStyles } from "./Style.js";
 
 export default function Widget() {
   const classes = useStyles();
