+++ src/components/Dashboard/Widgets/Filters.jsx
@@ -1,64 +1,21 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
 // import Avatar from "@material-ui/core/Avatar";
-import { grey } from "@material-ui/core/colors";
 // import Button from "@material-ui/core/Button";
 // import ghost from "../../../Assets/Ghost.jpg";
 
 import InputLabel from "@material-ui/core/InputLabel";
 import FormControl from "@material-ui/core/FormControl";
 import Select from "@material-ui/core/Select";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    maxWidth: 320,
-    marginBottom: "15px",
-    borderRadius: "20px",
-    width: "300px",
-  },
-  buttons: {
-    display: "flex",
-    flexWrap: "nowrap",
-    flexDirection: "column",
-    justifyContent: "space-evenly",
-    height: "fit-content",
-    margin: "10px",
-  },
-  button: {
-    backgroundColor: "#3AAFA9",
-    color: "white",
-    margin: "0 25px",
-    "&:hover": {
-      backgroundColor: "#27938d",
-    },
-  },
-  avatar: {
-    backgroundColor: grey[500],
-    width: "30px",
-    height: "30px",
-  },
-  header: {
-    backgroundColor: "#3AAFA9",
-    color: "white",
-    height: "15px",
-  },
-  formControl: {
-    margin: theme.spacing(1),
-    minWidth: 120,
-  },
-  selectEmpty: {
-    marginTop: theme.spacing(2),
-  },
-}));
+import { useStylesFilter } from "./Style.js";
 
 export default function Filters() {
-  const classes = useStyles();
+  const classes = useStylesFilter();
 
   const [state, setState] = React.useState({
     age: "",
-    name: "hai",
+    name: "hi",
   });
 
   const handleChange = (event) => {
@@ -74,10 +31,10 @@ export default function Filters() {
       <CardHeader
         // avatar={<Avatar src={ghost} className={classes.avatar} />}
         title="Filters"
-        className={classes.header}
+        className={classes.fheader}
       />
-      <div className={classes.buttons}>
-        <FormControl variant="outlined" className={classes.formControl}>
+      <div className={classes.fbuttons}>
+        <FormControl variant="outlined" className={classes.fformControl}>
           <InputLabel htmlFor="outlined-age-native-simple">Catagory</InputLabel>
           <Select
             native
@@ -95,7 +52,7 @@ export default function Filters() {
             <option value={30}>Other</option>
           </Select>
         </FormControl>
-        <FormControl variant="outlined" className={classes.formControl}>
+        <FormControl variant="outlined" className={classes.fformControl}>
           <InputLabel htmlFor="outlined-age-native-simple">City</InputLabel>
           <Select
             native
@@ -113,7 +70,7 @@ export default function Filters() {
             <option value={30}>Basrah</option>
           </Select>
         </FormControl>
-        <FormControl variant="outlined" className={classes.formControl}>
+        <FormControl variant="outlined" className={classes.fformControl}>
           <InputLabel htmlFor="outlined-age-native-simple">Color</InputLabel>
           <Select
             native
