+++ src/components/Dashboard/Widgets/LastPosts.jsx
@@ -1,102 +1,20 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
 // import Avatar from "@material-ui/core/Avatar";
-import { grey } from "@material-ui/core/colors";
 // import ghost from "../../../Assets/Ghost.jpg";
 import CardMedia from "@material-ui/core/CardMedia";
-
-import { withStyles } from "@material-ui/core/styles";
-import MuiAccordion from "@material-ui/core/Accordion";
-import MuiAccordionSummary from "@material-ui/core/AccordionSummary";
-import MuiAccordionDetails from "@material-ui/core/AccordionDetails";
 import Typography from "@material-ui/core/Typography";
 
-const Accordion = withStyles({
-  root: {
-    border: "1px solid rgba(0, 0, 0, .125)",
-    boxShadow: "none",
-    "&:not(:last-child)": {
-      borderBottom: 0,
-    },
-    "&:before": {
-      display: "none",
-    },
-    "&$expanded": {
-      margin: "auto",
-    },
-  },
-  expanded: {},
-})(MuiAccordion);
-
-const AccordionSummary = withStyles({
-  root: {
-    backgroundColor: "rgba(0, 0, 0, .03)",
-    borderBottom: "1px solid rgba(0, 0, 0, .125)",
-    marginBottom: -1,
-    minHeight: 56,
-    "&$expanded": {
-      minHeight: 56,
-    },
-  },
-  content: {
-    "&$expanded": {
-      margin: "12px 0",
-    },
-  },
-  expanded: {},
-})(MuiAccordionSummary);
-
-const AccordionDetails = withStyles((theme) => ({
-  root: {
-    padding: theme.spacing(2),
-  },
-}))(MuiAccordionDetails);
-
-const useStyles = makeStyles(() => ({
-  root: {
-    maxWidth: 320,
-    marginBottom: "15px",
-    borderRadius: "20px",
-    width: "300px",
-  },
-  buttons: {
-    display: "flex",
-    flexWrap: "nowrap",
-    flexDirection: "column",
-    justifyContent: "space-evenly",
-    height: "170px",
-  },
-  button: {
-    backgroundColor: "#3AAFA9",
-    color: "white",
-    margin: "0 25px",
-    "&:hover": {
-      backgroundColor: "#27938d",
-    },
-  },
-  avatar: {
-    backgroundColor: grey[500],
-    width: "30px",
-    height: "30px",
-  },
-  header: {
-    backgroundColor: "#3AAFA9",
-    color: "white",
-    height: "15px",
-  },
-  media: {
-    height: "0%",
-    paddingTop: "56.25%", // 16:9
-    marginTop: "0px",
-    marginBottom: "10px",
-    borderRadius: "10px",
-  },
-}));
+import {
+  Accordion,
+  AccordionSummary,
+  AccordionDetails,
+  useStylesLastPosts,
+} from "./Style.js";
 
 export default function LastPosts() {
-  const classes = useStyles();
+  const classes = useStylesLastPosts();
   const [expanded, setExpanded] = React.useState("panel1");
 
   const handleChange = (panel) => (event, newExpanded) => {
